refactor: 중복 코드 제거와 패키지 이동 #22
