--- src/main/java/com/munecting/api/global/auth/oidc/publicKey/client/PublicKeyClient.java
@@ -1,7 +1,0 @@
-package com.munecting.api.global.auth.oidc.publicKey.client;
-
-import com.munecting.api.global.auth.oidc.publicKey.OidcPublicKeyList;
-
-public interface PublicKeyClient {
-    OidcPublicKeyList getPublicKeys();
-}
